app/code/requirements_ummvar.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
eo-metadata-tools-cmr @ https://github.com/nasa/eo-metadata-tools/releases/download/latest-master/eo_metadata_tools_cmr-0.0.1-py3-none-any.whl
Flask==2.2.2
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
